requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
pytest==7.2.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
